# be/openapi.yaml
openapi: 3.0.3
info:
  title: cocono-chat backend API
  version: 0.1.0
  description: |
    REST API for cocono-chat (milestones 1–2: accounts + multi-device).

    ## Encoding conventions

    - All binary material (public keys, AES keys, signatures, nonces) is encoded as
      **base64url** (RFC 4648 §5, no padding).
    - Ed25519 public keys are the **raw 32-byte** key (`crypto.subtle.exportKey('raw')`
      in the browser).
    - Signatures are computed over the UTF-8 bytes of the **canonical JSON** of the
      signed object: object keys sorted lexicographically, no whitespace
      (e.g. `canonical({u,p,a,d,t})` → `{"a":...,"d":...,"p":...,"t":...,"u":...}`).
    - Signed payloads (signup, device enrollment) include a client timestamp `t`
      (epoch seconds) that must be within `SIGNED_PAYLOAD_MAX_AGE_SEC` (default 5
      minutes); each signature is additionally de-duplicated in Redis, so captured
      payloads cannot be replayed.

    ## Authentication

    There are no passwords. Login is a challenge-response proof of possession of a
    registered device's Ed25519 private key:

    1. `POST /api/auth/challenge` → one-time nonce `n` (Redis-backed, TTL, single-use).
       A nonce is issued for ANY username/device pair — unknown pairs get a nonce that
       simply never verifies, so the endpoint does not reveal which accounts exist.
    2. Client signs the nonce string (UTF-8 bytes) with the device key.
...
       `Authorization: Bearer <token>`.

...

...

...

...
servers:
...
paths:
...
  /api/users/{username}/stats:
    get:
      tags: [accounts]
      operationId: userStats
      summary: Reputation counts for a profile (numbers only, never identities)
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Exclusive vouch-stage buckets (addedBy / verifiedBy / trustedBy) plus the CoCo score (verifiedBy + 3×trustedBy, see docs/COCO_SCORE.md).
        '404': { description: 'No such user.' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT issued by /api/auth/verify.
...